Обновление
Добавлено новые функции для объекта Полилиния:
- поиск точки/точек пересечения с другими объектами;
- длина полилинии/полилиний;
- сгладить полилинию/полилинии;
- узнать угол между тремя точками полилинии;
- повернуть полилинию на новую позицию (курсор мыши);
- переместить полилинию на новую позицию (курсор мыши);
- создание плоскости (Полигона), при условии, что полилиния содержит больше 3-х точек;
- узнать в какую сторону построенная полилиния (по часовой или против часовой стрелки);
- комбинация клавиш "Shift+T" очищает выделенные объекты (но НЕ выходит из режима Selection).

Как работает?
Аналогично методам Линии.

Исправлено:
- прямоугольник выделения объектов -> теперь можно (в режиме Selection) зажатую кнопку мыши перемещать в любом направлении;
- ошибки предыдущих функций;
- рерайтинг кода.

+++ ShapesEditor2D/Models/Matrix.cs
@@ -9,7 +9,7 @@
 			_matrix = matrix;
 		}
 
-		public static Matrix CreateTranslationMatrix(double x, double y)
+		public static Matrix CreateTranslationMatrix(int x, int y)
 		{
 			double[,] translationMatrix = new double[,]
 			{
